+++ example/lib/main.dart
@@ -6,7 +6,8 @@ Future<void> main() async {
   Mirrorfly.init(
       baseUrl: 'https://api-uikit-qa.contus.us/api/v1/',
       licenseKey: 'your license key',
-      iOSContainerID: 'your container id');//Must be same as AppGroups given in Xcode Capability
+      iOSContainerID:
+          'your container id'); //Must be same as AppGroups given in Xcode Capability
   runApp(const MyApp());
 }