+++ lib/mirrorflychat.dart
@@ -1,3 +1,4 @@
+/// A MirrorFly Plugin to Experience an outstanding real time messaging solution.
 library mirrorfly_plugin;
 
 export 'fly_chat_method_channel.dart';
@@ -18,5 +19,6 @@ export 'model/export_model.dart';
 export 'model/call_constants.dart';
 export 'model/topic_metadata.dart';
 export 'model/topic_list.dart';
+export 'model/available_features.dart';
 export 'builder.dart';
 export 'flychat.dart';