+++ lib/model/call_constants.dart
@@ -1,4 +1,4 @@
-class CallAction{
+class CallAction {
   static const String inviteUsers = "INVITE_USERS";
   static const String answerCall = "ANSWER_CALL";
   static const String denyCall = "DENY_CALL";
@@ -14,9 +14,12 @@ class CallAction{
   static const String remoteVideoPaused = "REMOTE_VIDEO_PAUSED";
   static const String remoteVideoResumed = "REMOTE_VIDEO_RESUMED";
   static const String changeToAudioCall = "CHANGE_TO_AUDIO_CALL";
-  static const String videoCallConversionCancel = "ACTION_VIDEO_CALL_CANCEL_CONVERSION";
-  static const String videoCallConversionAccepted = "ACTION_VIDEO_CALL_CONVERSION_ACCEPTED";
-  static const String videoCallConversionRejected = "ACTION_VIDEO_CALL_CONVERSION_REJECTED";
+  static const String videoCallConversionCancel =
+      "ACTION_VIDEO_CALL_CANCEL_CONVERSION";
+  static const String videoCallConversionAccepted =
+      "ACTION_VIDEO_CALL_CONVERSION_ACCEPTED";
+  static const String videoCallConversionRejected =
+      "ACTION_VIDEO_CALL_CONVERSION_REJECTED";
   static const String remoteVideoAdded = "REMOTE_VIDEO_ADDED";
   static const String audioDeviceChanged = "AUDIO_DEVICE_CHANGED";
   static const String cameraSwitchSuccess = "CAMERA_SWITCH_SUCCESS";
@@ -29,19 +32,19 @@ class CallAction{
   static const String closeServerConnection = "ACTION_CLOSE_SERVER_CONNECTION";
 }
 
-class MuteStatus{
+class MuteStatus {
   static const String remoteAudioUnMute = "REMOTE_AUDIO_UN_MUTE";
   static const String remoteAudioMute = "REMOTE_AUDIO_MUTE";
   static const String remoteVideoMute = "REMOTE_VIDEO_MUTE";
   static const String remoteVideoUnMute = "REMOTE_VIDEO_UN_MUTE";
 }
 
-class CallType{
+class CallType {
   static const String oneToOne = "OneToOne";
   static const String groupCall = "GroupCall";
 }
 
-class CallStatus{
+class CallStatus {
   static const String connecting = "Connecting";
   static const String ringing = "Ringing";
   static const String attended = "Attended";
@@ -57,10 +60,11 @@ class CallStatus{
   static const String reconnected = "Reconnected";
   static const String calling = "Trying to Connect";
   static const String calling10s = "Calling... \n Trying to Connect";
-  static const String callingAfter10s = "User Seems to be Offline, Trying to Connect";
+  static const String callingAfter10s =
+      "User Seems to be Offline, Trying to Connect";
 }
 
-class AudioDeviceType  {
+class AudioDeviceType {
   static const String none = "none";
   static const String receiver = "receiver";
   static const String speaker = "speaker";